# lib/api-zod/src/generated/types/createGlossaryTerm.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Wiki API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGlossaryTerm {
  term: string;
  definition: string;
  synonyms?: string[];
  abbreviation?: string;
  nodeId?: string;
}