# lib/api-zod/src/generated/types/glossaryTerm.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Wiki API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GlossaryTerm {
  id: string;
  term: string;
  slug: string;
  definition: string;
  synonyms?: string[] | null;
  abbreviation?: string | null;
  nodeId?: string | null;
  createdBy?: string | null;
  createdAt?: Date;
  updatedAt?: Date;
}