# lib/api-zod/src/generated/types/searchAnalytics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Wiki API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SearchAnalyticsPeriod } from "./searchAnalyticsPeriod";
import type { SearchAnalyticsPopularQueriesItem } from "./searchAnalyticsPopularQueriesItem";
import type { SearchAnalyticsZeroResultQueriesItem } from "./searchAnalyticsZeroResultQueriesItem";

export interface SearchAnalytics {
  period?: SearchAnalyticsPeriod;
  totalSearches?: number;
  popularQueries?: SearchAnalyticsPopularQueriesItem[];
  zeroResultQueries?: SearchAnalyticsZeroResultQueriesItem[];
}