# lib/api-zod/src/generated/types/searchResultItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Wiki API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SearchResultItem {
  id: string;
  title: string;
  displayCode?: string | null;
  templateType: string;
  status: string;
  ownerId?: string | null;
  parentNodeId?: string | null;
  createdAt?: Date;
  updatedAt?: Date;
  rank?: number;
  headline?: string;
}