# lib/api-zod/src/generated/types/searchSuggestions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Wiki API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SearchSuggestionsAliasesItem } from "./searchSuggestionsAliasesItem";
import type { SearchSuggestionsNodesItem } from "./searchSuggestionsNodesItem";

export interface SearchSuggestions {
  nodes?: SearchSuggestionsNodesItem[];
  aliases?: SearchSuggestionsAliasesItem[];
}